# Refactor FileManager collation for easier patching