This PR extracts the collation related code from FileChooser and FileManager, to make it easier to add new sorting algorithms through user patches.

I also noticed a slight difference between which items got sorted depending on whether the `collate_mixed` option was set. Since the collation related code is now 'further away', I attempted to make it more consistent and predictable.

As an example, this is the user patch that I have been using: https://gist.github.com/wfdewith/978e4f8e09f481a480cb1bf645e8d518

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/koreader/11150)
<!-- Reviewable:end -->