@NiLuJe 2023-11-23
I wouldn't mind benchmarks to see if this has any measurable impact (because this codepath is surprisingly hot in my experience).

Also, ping @hius07 who worked on this more recently.