@wfdewith 2023-11-23
> I wouldn't mind benchmarks to see if this has any measurable impact (because this codepath is surprisingly hot in my experience and/or I/O makes it weird).

I'm going to look into this. The current sorting process is kind of wasteful, because it iterates three times over all entries in a directory. If you want to, I can attempt to reduce that number and see if it actually matters.

Regarding benchmarking, I assume you would do something like the following?
```lua
local start = timer.current()
-- do something expensive
print(timer.current() - start)
```
Or do you have fancier profiling?