@NiLuJe 2023-11-24
Another possibly relevant but generic and random comment: if the main goal is extensibility, one could also approach this by simply adding a new API call to register custom collate functions, depending on what makes the most sense re: the data structure of how everything is actually stored.