@NiLuJe 2023-11-24
> Regarding benchmarking

Unless you want to get extra fancy, yeah, something based on real time would be good enough, e.g., https://github.com/koreader/koreader/blob/bf03f40ef218797cc1288efaaac817ee479d1e3a/frontend/document/credocument.lua#L787-L791