@hius07 2023-11-24
I like the idea.
I'm not sure if it requires a separate module, maybe just add directly to FileChooser a hash table with collates, and an array of ids for menu order.