@wfdewith 2023-11-24
@hius07
> I like the idea. I'm not sure if it requires a separate module, maybe just add directly to FileChooser a hash table with collates, and an array of ids for menu order.

Fair enough, FileChooser and FileManager are already very tightly coupled and if the complexity can be reduced to a single table, I don't feel a strong need to keep the separate module.

---

I did some benchmarks with a corpus of 1000 random Project Gutenberg epubs. The file names all are `pgXXXX.epub`, where `XXXX` is a number. The epubs were distributed as follows:
* 10% in progress (with random percentage finished)
* 40% read
* 50% unread (no sidecar exists)

The tests were performed on my Kobo Clara HD.

Current situation:
```
11/24/23-14:16:52 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  5989.000 ms
11/24/23-14:16:58 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with strcoll took  1648.000 ms
11/24/23-14:17:02 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with natural took  1349.000 ms
11/24/23-14:17:07 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with access took  1227.000 ms
11/24/23-14:17:10 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with date took  1255.000 ms
11/24/23-14:17:12 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with size took  1215.000 ms
11/24/23-14:17:14 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with type took  1237.000 ms
11/24/23-14:17:19 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  3929.000 ms
11/24/23-14:17:25 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_last took  4607.000 ms
11/24/23-14:17:29 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  3230.000 ms
```

This PR:
```
11/24/23-14:27:50 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  6965.000 ms
11/24/23-14:27:58 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with strcoll took  1313.000 ms
11/24/23-14:28:01 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with natural took  1214.000 ms
11/24/23-14:28:03 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with access took  1224.000 ms
11/24/23-14:28:05 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with date took  1251.000 ms
11/24/23-14:28:08 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with size took  1201.000 ms
11/24/23-14:28:10 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with type took  1233.000 ms
11/24/23-14:28:14 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  3626.000 ms
11/24/23-14:28:21 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_last took  4976.000 ms
11/24/23-14:28:33 INFO  FileChooser:genItemTableFromPath: Sorting /mnt/onboard/.corpus with percent_unopened_first took  3147.000 ms
```

Honestly, it is really hard to say something definitively about these numbers, because filesystem caching and whatever the JIT cooks up probably play a large role in the differences. For example, for some reason, `natural` outperforms `strcoll`, which does not make sense to me at all.
The `percent_unopened_*` collates are much slower since they need to read the sidecars. The first `percentage_unopened_first` sort should warms up the filesystem caches for both the epubs and their sidecars. However, I cannot figure out why `percent_unopened_last` is so much slower than `percentage_unopened_first`.

Looking further at the time spent in various stages (in this PR):
```
11/24/23-15:03:28 INFO  FileChooser(percent_unopened_first): Walking directory /mnt/onboard/.corpus took  6007.000 ms
11/24/23-15:03:28 INFO  FileChooser(percent_unopened_first): Sorting items took    13.000 ms
11/24/23-15:03:28 INFO  FileChooser(percent_unopened_first): Creating menu items took     6.000 ms
11/24/23-15:03:35 INFO  FileChooser(strcoll): Walking directory /mnt/onboard/.corpus took  1683.000 ms
11/24/23-15:03:36 INFO  FileChooser(strcoll): Sorting items took    14.000 ms
11/24/23-15:03:36 INFO  FileChooser(strcoll): Creating menu items took    23.000 ms
11/24/23-15:03:38 INFO  FileChooser(natural): Walking directory /mnt/onboard/.corpus took  1343.000 ms
11/24/23-15:03:38 INFO  FileChooser(natural): Sorting items took    54.000 ms
11/24/23-15:03:38 INFO  FileChooser(natural): Creating menu items took    14.000 ms
11/24/23-15:03:41 INFO  FileChooser(access): Walking directory /mnt/onboard/.corpus took  1209.000 ms
11/24/23-15:03:41 INFO  FileChooser(access): Sorting items took     6.000 ms
11/24/23-15:03:41 INFO  FileChooser(access): Creating menu items took    22.000 ms
11/24/23-15:03:43 INFO  FileChooser(date): Walking directory /mnt/onboard/.corpus took  1242.000 ms
11/24/23-15:03:43 INFO  FileChooser(date): Sorting items took     9.000 ms
11/24/23-15:03:43 INFO  FileChooser(date): Creating menu items took    23.000 ms
11/24/23-15:03:45 INFO  FileChooser(size): Walking directory /mnt/onboard/.corpus took  1215.000 ms
11/24/23-15:03:45 INFO  FileChooser(size): Sorting items took     7.000 ms
11/24/23-15:03:45 INFO  FileChooser(size): Creating menu items took    12.000 ms
11/24/23-15:03:47 INFO  FileChooser(type): Walking directory /mnt/onboard/.corpus took  1252.000 ms
11/24/23-15:03:47 INFO  FileChooser(type): Sorting items took    14.000 ms
11/24/23-15:03:47 INFO  FileChooser(type): Creating menu items took    12.000 ms
11/24/23-15:03:51 INFO  FileChooser(percent_unopened_first): Walking directory /mnt/onboard/.corpus took  3428.000 ms
11/24/23-15:03:51 INFO  FileChooser(percent_unopened_first): Sorting items took    13.000 ms
11/24/23-15:03:51 INFO  FileChooser(percent_unopened_first): Creating menu items took     6.000 ms
11/24/23-15:03:57 INFO  FileChooser(percent_unopened_last): Walking directory /mnt/onboard/.corpus took  4854.000 ms
11/24/23-15:03:57 INFO  FileChooser(percent_unopened_last): Sorting items took    13.000 ms
11/24/23-15:03:57 INFO  FileChooser(percent_unopened_last): Creating menu items took     5.000 ms
11/24/23-15:04:02 INFO  FileChooser(percent_unopened_first): Walking directory /mnt/onboard/.corpus took  3196.000 ms
11/24/23-15:04:02 INFO  FileChooser(percent_unopened_first): Sorting items took    13.000 ms
11/24/23-15:04:02 INFO  FileChooser(percent_unopened_first): Creating menu items took     6.000 ms
11/24/23-15:04:07 INFO  FileChooser(percent_unopened_last): Walking directory /mnt/onboard/.corpus took  4067.000 ms
11/24/23-15:04:07 INFO  FileChooser(percent_unopened_last): Sorting items took    13.000 ms
11/24/23-15:04:07 INFO  FileChooser(percent_unopened_last): Creating menu items took     5.000 ms
```

It appears that sorting is almost irrelevant compared to the directory walk (`FileChooser:getList`), but the only thing this PR changed in that part of the process is the `item_func`. Optimizing filesystem interaction in general might be possible, for example by using `fstatat` instead of `stat` to get file information, but that would require patching LuaFileSystem and I very much doubt it actually makes a difference. I might do a bit more research into optimization, but not for this PR.

To conclude, despite the haphazardness of these benchmarks, this PR seems to exhibit the same behavior as the current situation. I'm fairly confident that it does not really introduce a significant performance penalty.

---

@NiLuJe 
> Another possibly relevant but mainly generic and random comment: if the main goal is extensibility, one could also approach this by simply adding a new API call to register custom collation functions, depending on what makes the most sense re: the data structure of how everything is actually stored.
>
> (As in, you don't necessarily need to make the data itself approachable and open, if it hampers readability or performance or whatever, it might make more sense to just provide a way to feed new data, possibly in a slightly different structure).

Maybe, but performance doesn't appear to matter, so I would rather keep it as straightforward as possible. Maintaining a code path that might only be used in user patches seems ill-advised.

---

Side note, I found that when KOReader is first started, the file list is sorted three times, though this is starting to get a bit out of scope of this PR.
1. [`FileChooser:init`](https://github.com/koreader/koreader/blob/bf03f40ef218797cc1288efaaac817ee479d1e3a/frontend/ui/widget/filechooser.lua#L103)
2. [`Menu:init`](https://github.com/koreader/koreader/blob/bf03f40ef218797cc1288efaaac817ee479d1e3a/frontend/ui/widget/menu.lua#L940) (this one is really smelly, why would the Menu class need to do this?)
3. [`CoverBrowser:refreshFileManagerInstance`](https://github.com/koreader/koreader/blob/bf03f40ef218797cc1288efaaac817ee479d1e3a/plugins/coverbrowser.koplugin/main.lua#L406-L418) (this one obviously only triggers when you have the cover browser enabled)

Another side note, sorting by access time does not work on my Kobo Clara HD, as the storage is mounted with `noatime`. Since FAT does not support `atime` in general and for other filesystems most eReaders probably configure `noatime` to increase storage performance and life expectancy, I suspect there are very few situations in which sorting by access time will work. Is this something that should just be caveat emptor, do we want to mark this on a device level, or do we simply ditch sorting by access time in general?